requirements.txt
blinker==1.9.0
cachelib==0.9.0
cffi==1.17.1
circuitbreaker==2.0.0
click==8.1.7
cryptography==44.0.0
Deprecated==1.2.15
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
mysql-connector-python==9.1.0
ordered-set==4.1.0
packaging==24.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.2
wrapt==1.17.0
